navbar and topnav restyling
