+++ public/index.html
@@ -14,21 +14,6 @@
       name="description"
       content="TV24 Africa News - Political, Economic, Development and African Leisure Report"
     />
-    <!-- <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" /> -->
-    <!--
-      manifest.json provides metadata used when your web app is installed on a
-      user's mobile device or desktop. See https://developers.google.com/web/fundamentals/web-app-manifest/
-    -->
-    <!-- <link rel="manifest" href="%PUBLIC_URL%/manifest.json" /> -->
-    <!--
-      Notice the use of %PUBLIC_URL% in the tags above.
-      It will be replaced with the URL of the `public` folder during the build.
-      Only files inside the `public` folder can be referenced from the HTML.
-
-      Unlike "/favicon.ico" or "favicon.ico", "%PUBLIC_URL%/favicon.ico" will
-      work correctly both with client-side routing and a non-root public URL.
-      Learn how to configure a non-root public URL by running `npm run build`.
-    -->
     <!-- bootstrap cdn -->
     <link
       rel="stylesheet"
@@ -54,6 +39,8 @@
       rel="stylesheet"
     />
     <link rel="preconnect" href="https://fonts.gstatic.com" />
+    <link href="http://fonts.cdnfonts.com/css/nimbussanl" rel="stylesheet" />
+
     <!-- Font awesome cdn -->
     <script
       src="https://kit.fontawesome.com/bb74a7de63.js"
