+++ src/App.css
@@ -1,21 +1,29 @@
+@import url('http://fonts.cdnfonts.com/css/nimbussanl');
 * {
   margin: 0;
   padding: 0;
+  box-sizing: border-box;
   -ms-box-sizing: border-box;
   -moz-box-sizing: border-box;
-  box-sizing: border-box;
 }
-
-.App {
-  margin: 0 auto;
-}
-
 body {
   overflow-x: hidden;
   overflow-y: scroll;
   scroll-behavior: smooth;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  font-family: 'NimbusSanL', sans-serif;
+  font-family: 'NimbusSanLCon', sans-serif;
+  font-size: clamp(0.688rem, 2vw, 0.8rem);
 }
-
 a:hover {
   text-decoration: none;
+}
+ul {
+  list-style: none;
+  margin: 0;
+  padding: 0;
+}
+.app {
+  margin: 0;
 }
