+++ src/components/generalNews/News.js
@@ -1,25 +1,13 @@
-import React, { useEffect, Fragment, useContext, useState } from 'react'
+import React, { useEffect, Fragment, useContext } from 'react'
 // import CommentForm from './CommentForm'
 // import ShareNews from './ShareNews'
 import bannerAds from './../../assets/images/bannerads.png'
-import ReactHtmlParser, {
-  processNodes,
-  convertNodeToElement,
-} from 'react-html-parser'
-// import { marked } from 'marked'
+import ReactHtmlParser from 'react-html-parser'
 import Nav from '../reusables/navigation/Nav/nav'
 import Footer from '../reusables/navigation/Footer/footer'
 import { useParams, Link } from 'react-router-dom'
-import { Button } from 'react-bootstrap'
-
-// import {
-//   getNewsComments,
-//   getNewsFeed,
-//   getSingleNews,
-// } from '../../context/news/NewsApi'
 import Loader from '../loader/Loader'
 import './allNews.css'
-// import NewsComments from './NewsComments'
 import { NotLoggedIn, LoggedInNotSubscribed } from './FreeReaderPersuader'
 // import { PopulateReadersList } from '../homepage/politics/ReaderList'
 // import { ContactsAds1 } from '../ContactUs/mainSection/ContactsAds'
@@ -35,88 +23,21 @@ import Moment from 'react-moment'
 import { HtmlParseOptions } from '../../_helper/parseNewsHtml'
 import { useViewPort } from '../../components/hooks/Viewport'
 import '../category/newscategory.css'
-
-function transform(node, index) {
-  if (node.type === 'tag' && node.name === 'span') {
-    return null
-  }
-  if (node.type === 'tag' && node.name === 'ul') {
-    node.name = 'ol'
-    return convertNodeToElement(node, index, transform)
-  }
-
-  if (node.type === 'tag' && node.name === 'b') {
-    return <i key={index}>{processNodes(node.children, transform)}</i>
-  }
-  if (node.type === 'tag' && node.name === 'a') {
-    node.attribs.target = '_blank'
-
-    return convertNodeToElement(node, index, transform)
-  }
-
-  if (node.type === 'tag' && node.name === 'button') {
-    return (
-      <Button variant="contained" color="primary" key={index}>
-        {processNodes(node.children, transform)}
-      </Button>
-    )
-  }
-}
-
-const options = {
-  decodeEntities: true,
-  transform,
-}
+import NewsComments from './NewsComments'
+import NewsForm from './NewsForm'
+import RelatedNews from './RelatedNews'
+import ShareNews from './ShareNews'
 
 const GetNews = () => {
-  const [comments, setComments] = useState({
-    comment: '',
-    name: '',
-    email: '',
-    website: '',
-  })
-  const [notifyComment, setNotifyComment] = useState(false)
-  const [commentLength, setCommentLength] = useState(3)
-  const [saveEmail, setSaveEmail] = useState(false)
-  const [notifyPost, setNotifyPost] = useState(false)
   const contextComment = useContext(commentsContext)
   const userContext = useContext(authContext)
-  const { user, isAuthenticated, emailSub } = userContext
+  const { user, isAuthenticated } = userContext
   const newsFeedContext = useContext(newsContext)
   const { news, loading, singleNews, getNews, getSingleNews } = newsFeedContext
-  const {
-    comment_loading,
-    postComments,
-    getComments,
-    newsComment,
-  } = contextComment
-  const { comment, name, email, website } = comments
-  // const [comments, setComments] = useState(null)
+  const { comment_loading } = contextComment
   const { slug } = useParams()
   const { width } = useViewPort()
   const breakpoint = 1150
-
-  const onChange = (e) => {
-    setComments({ ...comments, [e.target.name]: e.target.value })
-  }
-
-  const onSubmit = (e) => {
-    e.preventDefault()
-    postComments(name, comment, singleNews[0].id, singleNews[0].post_title)
-    if (notifyPost || notifyComment) {
-      emailSub(email)
-    }
-    setComments({
-      comment: '',
-      name: '',
-      email: '',
-      website: '',
-    })
-  }
-
-  const handleComment = () => {
-    setCommentLength((prev) => prev + 2)
-  }
 
   const getAdjacentPosts = (slug) => {
     if (singleNews.length === 0) return ''
@@ -145,7 +66,6 @@ const GetNews = () => {
 
   useEffect(() => {
     getSingleNews(slug)
-    getComments(slug)
     //eslint-disable-next-line
   }, [slug])
 
@@ -154,18 +74,7 @@ const GetNews = () => {
     //eslint-disable-next-line
   }, [])
 
-  useEffect(() => {
-    if (user) {
-      setComments({
-        comment: '',
-        name: `${user.data.firstname} ${user.data.lastname}`,
-        email: `${user.data.email}`,
-        website: '',
-      })
-    }
-  }, [user])
-
-  if (loading || singleNews === null || news === null || newsComment === null) {
+  if (loading || singleNews === null || news === null || comment_loading) {
     return <Loader />
   }
 
@@ -224,7 +133,7 @@ const GetNews = () => {
                             ? singleNews[0].post_description
                             : singleNews[0].post_description.substring(0, 1500)
                         }`,
-                        options,
+                        HtmlParseOptions,
                       )}
                     </div>
                     {singleNews[0].post_type === 'premium' &&
@@ -249,45 +158,7 @@ const GetNews = () => {
                         ''
                       )}
                     </div>
-
-                    <div className="next-prev-section">
-                      <div className="news-social-icons">
-                        <p className="share">Share this story</p>
-                        <span className="news-social-icons-items n-facebook">
-                          <i className="fab fa-facebook"></i>
-                        </span>
-                        <span className="news-social-icons-items n-twitter">
-                          <i className="fab fa-twitter"></i>
-                        </span>
-                        <span className="news-social-icons-items n-instagram">
-                          <i className="fab fa-instagram"></i>
-                        </span>
-                        <span className="news-social-icons-items n-whatsapp">
-                          <i className="fab fa-whatsapp"></i>
-                        </span>
-                        <span className="news-social-icons-items n-linkedin">
-                          <i className="fab fa-linkedin"></i>
-                        </span>
-                      </div>
-                      <div className="next-or-prev-section">
-                        <div className="previous">
-                          <p className="previous-article">Previous Article</p>
-                          <Link
-                            to={`/post/${previous.slug}`}
-                            className="prev-link"
-                          >
-                            {previous.slug}
-                          </Link>
-                        </div>
-                        <div className="next">
-                          <p className="next-article">Next Article</p>
-                          <Link to={`/post/${next.slug}`} className="next-link">
-                            {next.slug}
-                          </Link>
-                        </div>
-                      </div>
-                    </div>
-
+                    <ShareNews next={next} previous={previous} />
                     <div className="related-articles-section">
                       <button className="related-articles-btn">
                         related articles
@@ -297,163 +168,18 @@ const GetNews = () => {
                           {currentCategoryNewsWithoutSingleNews
                             .slice(0, 4)
                             .map((categ, idx) => (
-                              <Col className="related-card" key={categ.id}>
-                                <Link
-                                  to={`/post/${categ.slug}`}
-                                  className="related-link"
-                                >
-                                  <Card className="l-card">
-                                    <Card.Img
-                                      variant="top"
-                                      src={`https://api.tv24africa.com/public/storage/post_image/${categ.featured_image}`}
-                                      className="mb-3 card-img-related"
-                                    />
-                                    <p className="premium-badge-left">
-                                      {categ.post_type === 'premium'
-                                        ? `${categ.post_type}`
-                                        : ''}
-                                    </p>
-                                    <Card.Body className="l-card-body">
-                                      <Card.Text>{categ.slug}</Card.Text>
-                                    </Card.Body>
-                                  </Card>
-                                </Link>
-                              </Col>
+                              <RelatedNews
+                                key={categ.id}
+                                slug={categ.slug}
+                                featured_image={categ.featured_image}
+                                post_type={categ.post_type}
+                              />
                             ))}
                         </Row>
                       </div>
                     </div>
-
-                    {/* POST COMMENT && COMMENT SECTION */}
-                    <div className="comments-comments">
-                      <h5 className="comment-form-header">Comments</h5>
-                      {!comment_loading && newsComment.length === 0 ? (
-                        <h5>Be the first to comment on this post!</h5>
-                      ) : (
-                        newsComment
-                          .slice(0, commentLength)
-                          .map((eachComment) => (
-                            <div
-                              className="comments-comments-container"
-                              key={eachComment.id}
-                            >
-                              <div className="comment-details">
-                                <h5 className="comment-name">
-                                  {eachComment.name}
-                                </h5>
-                                <p className="comment-text">
-                                  {eachComment.comment}
-                                </p>
-                              </div>
-                              <p className="comment-date">
-                                <Moment format="MMMM Do YYYY">
-                                  {eachComment.created_at}
-                                </Moment>
-                              </p>
-                            </div>
-                          ))
-                      )}
-                      <div className="comment-load-more">
-                        <button className="load-more" onClick={handleComment}>
-                          Load More...
-                        </button>
-                      </div>
-                    </div>
-                    <div className="comment-form">
-                      <h5 className="comment-form-header">Leave a reply</h5>
-                      <form className="post-comment-form" onSubmit={onSubmit}>
-                        <div className="post-group">
-                          <textarea
-                            placeholder="Comment"
-                            className="form-control area-input"
-                            rows={8}
-                            name="comment"
-                            value={comment}
-                            onChange={onChange}
-                            required
-                          />
-                        </div>
-                        <div className="post-group">
-                          <input
-                            type="text"
-                            placeholder="Name"
-                            className="form-control post-input"
-                            name="name"
-                            value={name}
-                            onChange={onChange}
-                            required
-                          />
-                        </div>
-                        <div className="post-group">
-                          <input
-                            type="email"
-                            placeholder="Email"
-                            className="form-control post-input"
-                            name="email"
-                            value={email}
-                            onChange={onChange}
-                            required
-                          />
-                        </div>
-                        <div className="post-group">
-                          <input
-                            type="text"
-                            placeholder="Website"
-                            className="form-control post-input"
-                            name="website"
-                            value={website}
-                            onChange={onChange}
-                            required
-                          />
-                        </div>
-                        <div className="form-group check-section">
-                          <input
-                            type="checkbox"
-                            name="agree"
-                            className="agree-checkbox"
-                            checked={saveEmail}
-                            onChange={() => setSaveEmail(true)}
-                          />
-                          <label htmlFor="agree" className="reg-agree-label">
-                            Save my name, email, and website in this browser for
-                            the next time I comment
-                          </label>
-                        </div>
-                        <div className="form-group check-section">
-                          <input
-                            type="checkbox"
-                            name="agree"
-                            className="agree-checkbox"
-                            checked={notifyComment}
-                            onChange={() => setNotifyComment(true)}
-                          />
-                          <label htmlFor="agree" className="reg-agree-label">
-                            Notify me of follow-up comments by email
-                          </label>
-                        </div>
-                        <div className="form-group check-section">
-                          <input
-                            type="checkbox"
-                            name="agree"
-                            className="agree-checkbox"
-                            checked={notifyPost}
-                            onChange={() => setNotifyPost(true)}
-                          />
-                          <label htmlFor="agree" className="reg-agree-label ">
-                            Notify me of new posts by email
-                          </label>
-                        </div>
-                        <div className="post-group check-section">
-                          <button
-                            className="post-comment-btn contact-form-btn"
-                            type="submit"
-                            disabled={isAuthenticated}
-                          >
-                            {comment_loading ? 'Loading...' : 'Post Comment'}
-                          </button>
-                        </div>
-                      </form>
-                    </div>
+                    <NewsComments slug={slug} />
+                    <NewsForm singleNews={singleNews} />
                   </div>
                 )}
               </div>
