--- src/components/generalNews/NewsCard.js
@@ -1,22 +1,0 @@
-// import { Card } from "react-bootstrap";
-// const SportCard = ({ post_description, post_type, post_title, featured_image, id }) => {
-//   return (
-//     <div className="">
-//       <Card className="sprt-crd-mp">
-//         <Card.Img
-//           variant="top"
-//           src={`public/storage/${featured_image}/${id}`}
-//           className=""
-//           alt={`${post_type}`}
-//         />
-//         <Card.Body className="">
-//           <Card.Text>
-//             <p>{post_title}</p>
-//             <p>{post_description}</p>
-//           </Card.Text>
-//         </Card.Body>
-//       </Card>
-//     </div>
-//   );
-// };
-// export default SportCard;
