+++ src/components/generalNews/ShareNews.js
@@ -1,52 +1,43 @@
-// import React from "react";
-// import { Link } from "react-router-dom";
-// import "./allNews.css";
-// const ShareNews = (props) => {
-//   const { post_title, post_description } = props;
-//   return (
-//     <div className="row share-icons-container">
-//       <span className="small">Share this story</span>
-//       <div className="share-news">
-//         <a
-//           target="_blank"
-//           href={`https://twitter.com/intent/tweet/?text=${post_title}. Read More on&url=tv24africa.com`}
-//         >
-//           <span className="share-icon-wrap twt">
-//             <i className="fab fa-twitter"></i>{" "}
-//           </span>
-//         </a>
-//         <a
-//           target="_blank"
-//           href="https://facebook.com/sharer/sharer.php?u=tv24africa.com"
-//         >
-//           <span className="share-icon-wrap fb">
-//             <i className="fab fa-facebook"></i>{" "}
-//           </span>
-//         </a>
-//         <a target="_blank" hreh="">
-//           <span className="share-icon-wrap insta">
-//             <i className="fab fa-instagram"></i>{" "}
-//           </span>
-//         </a>
-//         <a
-//           target="_blank"
-//           href={`https://api.whatsapp.com/send?text=${post_title}. Read More on tv24africa.com`}
-//         >
-//           <span className="share-icon-wrap twt">
-//             <i className="fab fa-whatsapp wtsapp"></i>{" "}
-//           </span>
-//         </a>
-//         <a
-//           target="_blank"
-//           href={`https://www.linkedin.com/shareArticle?mini=true&url=https://tv24africa.com&title=${post_title} &summary=${post_title}&source=tv24africa`}
-//         >
-//           <span className="share-icon-wrap lnk">
-//             <i className="fab fa-linkedin"></i>{" "}
-//           </span>
-//         </a>
-//       </div>
-//     </div>
-//   );
-// };
+import React from 'react'
+import { Link } from 'react-router-dom'
 
-// export default ShareNews;
+const ShareNews = ({ next, previous }) => {
+  return (
+    <div className="next-prev-section">
+      <div className="news-social-icons">
+        <p className="share">Share this story</p>
+        <span className="news-social-icons-items n-facebook">
+          <i className="fab fa-facebook"></i>
+        </span>
+        <span className="news-social-icons-items n-twitter">
+          <i className="fab fa-twitter"></i>
+        </span>
+        <span className="news-social-icons-items n-instagram">
+          <i className="fab fa-instagram"></i>
+        </span>
+        <span className="news-social-icons-items n-whatsapp">
+          <i className="fab fa-whatsapp"></i>
+        </span>
+        <span className="news-social-icons-items n-linkedin">
+          <i className="fab fa-linkedin"></i>
+        </span>
+      </div>
+      <div className="next-or-prev-section">
+        <div className="previous">
+          <p className="previous-article">Previous Article</p>
+          <Link to={`/post/${previous.slug}`} className="prev-link">
+            {previous.slug}
+          </Link>
+        </div>
+        <div className="next">
+          <p className="next-article">Next Article</p>
+          <Link to={`/post/${next.slug}`} className="next-link">
+            {next.slug}
+          </Link>
+        </div>
+      </div>
+    </div>
+  )
+}
+
+export default ShareNews
