+++ src/components/reusables/header.css
@@ -49,7 +49,7 @@ ul {
   gap: 13px;
 }
 
-@media screen and (max-width: 1250px) {
+@media (max-width: 1250px) {
   .soc-nav {
     display: none;
   }
